backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
sqlmodel==0.0.38
alembic==1.18.4
asyncpg==0.31.0
aiosqlite==0.22.1
pydantic-settings==2.14.0
python-dotenv==1.2.2
python-multipart==0.0.27
passlib==1.7.4
PyJWT==2.12.1
httpx==0.28.1
SQLAlchemy==2.0.49
pydantic==2.13.3
typing_extensions==4.15.0
rich==15.0.0
extra-streamlit-components==-0.1.81
streamlit-cookies-manager==0.2.0